+++ package.json
@@ -14,16 +14,18 @@
     "url": "https://github.com/orzyyyy/memo"
   },
   "license": "MIT",
+  "author": "orzyyyy",
   "scripts": {
-    "analyse": "rimraf dist && cross-env BUILD_ENV=analyse nino koei --not-copy-assets",
-    "build": "rimraf dist && cross-env BUILD_ENV=prod nino koei --not-copy-assets",
-    "build:gh-pages": "rimraf dist && cross-env BUILD_ENV=prod PUBLISH_TO=github nino koei --not-copy-assets",
+    "analyse": "rimraf dist && cross-env BUILD_ENV=analyse nino koei",
+    "build": "rimraf dist && cross-env BUILD_ENV=netlify nino koei",
+    "build:business": "rimraf dist && cross-env BUILD_ENV=business nino koei",
+    "build:gh-pages": "rimraf dist && cross-env BUILD_ENV=gh-pages nino koei",
     "clean": "rimraf node_modules package-lock.json bin dist && npm i",
     "codecov": "nino test --codecov",
     "compile:server": "tsc -p server",
     "d": "npm run compile:server -- --watch",
     "deploy": "npm-run-all clean build:gh-pages pub",
-    "dev": "rimraf dist && nino koei -w -d --not-copy-assets",
+    "dev": "rimraf dist && cross-env BUILD_ENV=dev nino koei -w -d",
     "lint": "nino eslint",
     "lint-staged": "lint-staged",
     "prettier": "nino prettier",
